Move agentCheck and mapping to util
